Add files via upload
The code is written in the python. Where all the  define a class and write the code for insert the data into the Stack. Remove the element in the Stock. Access the peek element in the Stack. Find the Stack is empty and Stack is full both code are written in the file